server/requirements.txt
aiofile==3.8.8
aiohttp==3.9.5
aiosignal==1.3.1
aniso8601==9.0.1
async-lru==2.0.4
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.119
botocore==1.34.119
bs4==0.0.2
caio==0.9.17
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cognitojwt==1.4.1
cryptography==43.0.0
ecdsa==0.19.0
et-xmlfile==1.1.0
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
frozenlist==1.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
pywatchman==1.4.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
soupsieve==2.5
tzdata==2024.1
urllib3==1.26.19
Werkzeug==3.0.1
yarl==1.9.4
